+++ back/src/main/java/fr/centralesupelec/thuv/dtos/ContainerStatusDto.java
@@ -1,5 +1,5 @@
 package fr.centralesupelec.thuv.dtos;
 
 public enum ContainerStatusDto {
-    OK, KO
+    PENDING, CHECKING, OK, KO
 }