+++ back/src/main/java/fr/centralesupelec/thuv/dtos/ContainerSwarmStateDto.java
@@ -1,0 +1,17 @@
+package fr.centralesupelec.thuv.dtos;
+
+public enum ContainerSwarmStateDto {
+    NEW,
+    PENDING,
+    ASSIGNED,
+    ACCEPTED,
+    PREPARING,
+    READY,
+    STARTING,
+    RUNNING,
+    COMPLETE,
+    SHUTDOWN,
+    FAILED,
+    REJECTED,
+    ORPHANED,
+}