+++ back/src/main/java/fr/centralesupelec/thuv/web/student/ContainerController.java
@@ -19,6 +19,7 @@ import fr.centralesupelec.thuv.service.*;
 import fr.centralesupelec.thuv.storage.ShutdownStatusStorage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -117,7 +118,7 @@ public class ContainerController {
         return optionalContainer.orElse(null);
     }
 
-    @GetMapping(value = "logs/{courseId}")
+    @GetMapping(value = "logs/{courseId}", produces = MediaType.TEXT_PLAIN_VALUE)
     public String getLogs(
             @PathVariable("courseId") long courseId,
             @AuthenticationPrincipal(errorOnInvalidType = true) final MyUserDetails principal